expose data grid

+++ libmini/mini/minigeom.h
@@ -63,7 +63,7 @@ class minigeom_base
    miniv3d getpoint() const {return(pnt);}
    miniv3d getvector() const {return(vec);}
 
-   miniv3d getpoint(const float lambda) const {return(pnt+lambda*vec);}
+   miniv3d getpoint(const double lambda) const {return(pnt+lambda*vec);}
 
    miniv3d getminpoint() const {return(pnt+minlambda*vec);}
    miniv3d getmaxpoint() const {return(pnt+maxlambda*vec);}
